move goose models into its own component
