+++ ui/desktop/src/components/settings/models/ProviderButtons.tsx
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { Button } from '../../ui/button';
 import { Switch } from '../../ui/switch';
 import { useActiveKeys } from '../api_keys/ActiveKeysContext';
-import { model_docs_link, goose_models } from './hardcoded_stuff';
+import { model_docs_link } from './hardcoded_stuff';
+import { goose_models } from './goose_models';
 import { useModel } from './ModelContext';
 import { useHandleModelSelection } from './utils';
